test/python/requirements.txt
dana-python-api-client>=0.1.0
pytest>=7.2.1
pytest-cov>=2.8.1
